# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IT Asset Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: getServices
      tags: [services]
      summary: Get all service records
      security:
        - bearerAuth: []
      parameters:
        - name: assetId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Service records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceRecord"
    post:
      operationId: createService
      tags: [services]
      summary: Create service record
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "201":
          description: Service record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRecord"
  /services/{id}:
    put:
      operationId: updateService
      tags: [services]
      summary: Update service record
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceRequest"
      responses:
        "200":
          description: Service record updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRecord"
  /licenses:
    get:
      operationId: getLicenses
      tags: [licenses]
      summary: Get all software licenses
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [Available, Assigned, Expired]
      responses:
        "200":
          description: Licenses list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/License"
    post:
      operationId: createLicense
      tags: [licenses]
      summary: Create software license
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLicenseRequest"
      responses:
        "201":
          description: License created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
  /licenses/{id}:
    put:
      operationId: updateLicense
      tags: [licenses]
      summary: Update license
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLicenseRequest"
      responses:
        "200":
          description: License updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
    delete:
      operationId: deleteLicense
      tags: [licenses]
      summary: Delete license
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: License deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /assign-license:
    post:
      operationId: assignLicense
      tags: [licenses]
      summary: Assign license to employee
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignLicenseRequest"
      responses:
        "200":
          description: License assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseAssignment"
  /license-assignments:
    get:
      operationId: getLicenseAssignments
      tags: [licenses]
      summary: Get license assignments
      security:
        - bearerAuth: []
      parameters:
        - name: licenseId
          in: query
          schema:
            type: integer
        - name: employeeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: License assignments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LicenseAssignmentWithDetails"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: string
        name:
          type: string
        email:
          type: string
        department:
          type: string
        role:
          type: string
          enum: [admin, employee]
        isActive:
          type: boolean
        isLocked:
          type: boolean
        failedAttempts:
          type: integer
        lockedAt:
          type: string
          format: date-time
        doj:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - employeeId
        - name
        - role
        - isActive
        - isLocked
        - failedAttempts
...
    Asset:
      type: object
      properties:
        id:
          type: integer
        sn:
          type: string
        description:
          type: string
        category:
          type: string
        subCategory:
          type: string
        assetNumber:
          type: string
        owner:
          type: string
        employeeId:
          type: string
        custodian:
          type: string
        department:
          type: string
        email:
          type: string
        assignedDate:
          type: string
        location:
          type: string
        os:
          type: string
        ram:
          type: string
        rom:
          type: string
        osVersion:
          type: string
        serialNumber:
          type: string
        makeModel:
          type: string
        supplier:
          type: string
        invoice:
          type: string
        warranty:
          type: string
        status:
          type: string
          enum: [Available, Assigned, "In Service"]
        remarks:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - assetNumber
        - status
...
    ServiceRecord:
      type: object
      properties:
        id:
          type: integer
        assetId:
          type: integer
        serviceDate:
          type: string
        vendor:
          type: string
        cost:
          type: number
        status:
          type: string
          enum: [Pending, "In Progress", Completed, Cancelled]
        remarks:
          type: string
        createdAt:
          type: string
          format: date-time
        asset:
          $ref: "#/components/schemas/Asset"
      required:
        - id
        - assetId
        - status
    CreateServiceRequest:
      type: object
      properties:
        assetId:
          type: integer
        serviceDate:
          type: string
        vendor:
          type: string
        cost:
          type: number
        status:
          type: string
          enum: [Pending, "In Progress", Completed, Cancelled]
        remarks:
          type: string
      required:
        - assetId
        - status
    UpdateServiceRequest:
      type: object
      properties:
        serviceDate:
          type: string
        vendor:
          type: string
        cost:
          type: number
        status:
          type: string
          enum: [Pending, "In Progress", Completed, Cancelled]
        remarks:
          type: string
    License:
      type: object
      properties:
        id:
          type: integer
        licenseName:
          type: string
        licenseKey:
          type: string
        vendor:
          type: string
        purchaseDate:
          type: string
        expiryDate:
          type: string
        totalSeats:
          type: integer
        usedSeats:
          type: integer
        status:
          type: string
          enum: [Available, Assigned, Expired]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - licenseName
        - totalSeats
        - usedSeats
        - status
    CreateLicenseRequest:
      type: object
      properties:
        licenseName:
          type: string
        licenseKey:
          type: string
        vendor:
          type: string
        purchaseDate:
          type: string
        expiryDate:
          type: string
        totalSeats:
          type: integer
        status:
          type: string
          enum: [Available, Assigned, Expired]
      required:
        - licenseName
        - totalSeats
    UpdateLicenseRequest:
      type: object
      properties:
        licenseName:
          type: string
        licenseKey:
          type: string
        vendor:
          type: string
        purchaseDate:
          type: string
        expiryDate:
          type: string
        totalSeats:
          type: integer
        status:
          type: string
          enum: [Available, Assigned, Expired]
    LicenseAssignment:
      type: object
      properties:
        id:
          type: integer
        licenseId:
          type: integer
        employeeId:
          type: integer
        assignedDate:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - licenseId
        - employeeId
        - isActive
    LicenseAssignmentWithDetails:
      allOf:
        - $ref: "#/components/schemas/LicenseAssignment"
        - type: object
          properties:
            license:
              $ref: "#/components/schemas/License"
            employee:
              $ref: "#/components/schemas/Employee"
    AssignLicenseRequest:
      type: object
      properties:
        licenseId:
          type: integer
        employeeId:
          type: integer
        assignedDate:
          type: string
      required:
        - licenseId
        - employeeId
...