# lib/api-zod/src/generated/types/assignLicenseRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IT Asset Management API
 * OpenAPI spec version: 0.1.0
 */

export interface AssignLicenseRequest {
  licenseId: number;
  employeeId: number;
  assignedDate?: string;
}